move point of contact to own component
